+++ src/modules/index.ts
@@ -1,2 +1,5 @@
 export * from './sync_roadmap_ship_target_with_subtasks';
 export * from './deploy_preview_label_handler';
+export * from './sync_milestones_with_subtasks';
+export * from './deploy_notifications';
+export * from './engineering_projects_manager';
